requirements.txt
mechanize
requests_html
bs4
pymongo
flask